requirements.txt
fusepy==2.0.1
boto==2.6.0
iso8601==0.1.4
PyYAML==3.10
